Only run process tracking if not already set up
This block, bound to the SIGINT event, was running
when the watch task reloaded the app. This
replaced `.start.pid` with a new version
containing an updated pid while the watch task was
stopping the original process from terminating.

The result was when you finally tried to kill it
manually, the pid in `.start.pid` was for a new
process, but the old one was the one still
running.

+++ run_grunt.js
@@ -12,11 +12,14 @@ if (argv._.length) {
 } else {
   grunt.cli();
 }
-fs.writeFileSync(pidFile, process.pid, fileOptions);
-process.on('SIGINT', function() {
-  var pid = fs.readFileSync(pidFile, fileOptions);
 
-  fs.unlink(pidFile);
-  process.kill(pid, 'SIGTERM');
-  process.exit();
-});
+if (!fs.existsSync(pidFile)) {
+  fs.writeFileSync(pidFile, process.pid, fileOptions);
+  process.on('SIGINT', function() {
+    var pid = fs.readFileSync(pidFile, fileOptions);
+
+    fs.unlink(pidFile);
+    process.kill(pid, 'SIGTERM');
+    process.exit();
+  });
+}
